+++ RestaurantManagement/Models/ChiTietMon.cs
@@ -11,19 +11,19 @@ namespace QuanLyNhaHang.Models
     {
         private string _maMon;
         private string _tenNL;
-        private int _soluong;
+        private float _soluong;
 
         public string MaMon { get { return _maMon; } set { _maMon = value; OnPropertyChanged(); } }
         public string TenNL { get { return _tenNL; } set { _tenNL = value; OnPropertyChanged(); } }
-        public int SoLuong { get { return _soluong; } set { _soluong = value; OnPropertyChanged(); } }
-        public ChiTietMon(string TenNL, string MaMon, int SoLuong = 0)
+        public float SoLuong { get { return _soluong; } set { _soluong = value; OnPropertyChanged(); } }
+        public ChiTietMon(string TenNL, string MaMon, float SoLuong = 0)
         {
             _maMon = MaMon;
             _tenNL = TenNL;
             _soluong = SoLuong;
         }
         
-        public ChiTietMon(string TenNL, int SoLuong)
+        public ChiTietMon(string TenNL, float SoLuong)
         {
             _tenNL = TenNL;
             _soluong= SoLuong;
